+++ src/advanced/components/ProductPage.tsx
@@ -11,7 +11,7 @@ import {
 
 // ✅ Jotai hooks import
 import { useCartJotai } from "../hooks/useCart";
-import { useNotificationJotai } from "../utils/hooks/useNotification";
+import { useNotificationJotai } from "../hooks/useNotification";
 
 // ✅ 비즈니스 로직 import
 import {
